+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+###Fixed
+
+* Delete vote after poll ended
+
 ## 3.2.0
 
 ### Added